% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{calculate the bmi}
\usage{
calculate_bmi(weight_kg, height_cm)
}
\arguments{
\item{weight_kg}{numeric}

\item{height_cm}{numeric}
}
\value{
list with numeric BMI and category
\describe{
\item{bmi}{Numeric. BMI values are rounded to 1 decimal}
\item{category}{character. }
\code{'underwight'},
\code{'normal weight'}
\code{'overweight'}
\code{'obese'}
}
}
\description{
we are creating this function to derive the bmi with weight and height
}
\examples{
calculate_bmi(95,168)

}
